Make sure that coverage gets calculated properly
This changes it so that tox doesn't make a source distribution (sdist)
out of the repo (ptero_shell_command), but instead just points at the
local code.

We were losing the coverage because it was attributing it to places
like:

.tox/dev/lib/python2.7/site-packages/ptero_shell_command/api/v1/views.py

instead of:

ptero_shell_command/api/v1/views.py

This was only the case for the processes launched by honcho, the code
directly accessed in the unittests was being counted correctly, leading
to very low (but not zero) coverage percentages.