Revert "remove bash alternative setup"
Ubuntu changed to using Dash as /bin/sh:
https://wiki.ubuntu.com/DashAsBinSh

The problem is that /bin/bash exec <command> will replace the bash
process with the command that is launched (behavior we need).  The dash
documentation claims it will do that too, but it appears to be bugged.

What happens instead is that you have these /bin/sh processes in your
process tree.  Honcho gives them the SIGTERM and they immediately
terminate, which lets the Honcho process end.  Then, a short while
later, the underlying processes get the SIGTERM from our specialized
cleanup code and try to handle it.  The signal handler on the celery
workers tries to say something to stdout or stderr, but the PIPE is
broken, since the Honcho process ended.  So while they're trying to
handle SIGTERM, they run into this broken pipe and get hit with a
SIGPIPE.  This makes it impossible for these processes to be terminated
with any sequence of SIGTERM/SIGINTs. Finally our specialized cleanup
code sends everyone the SIGKILL and they die, without ever
writing out the .coverage.* files.

In the /bin/bash case, Honcho sends the SIGTERM directly to the
processes that are responsible for creating the .coverage.* files.  The
SIGTERM can be safely handled becuase the Honcho process is still alive.
After writing the .coverage.* files they exit, allowing the Honcho
process (which is waiting on them) to terminate.

This reverts commit e7742878adaf4d87f9f8d168895fab6bfc8955ee.