Silence environment substitution error messages
Even when the fork-worker testenv is not in use, the environment
substitution {env:} is still performed.  This commit adds an extra : to
the substitution expression so that an empty string is substituted when
CELERY_BROKER_URL or CELERY_BROKER_BACKEND are not set so that other
testenv's such as py27 (the default testenv used for testing) may work
normally.