Allow numbers in commandLine property
Because the JSON encoder in Perl cannot tell the difference between a
number and a string representing a number, it encodes it as a number.
This means that without this change, one cannot (from perl) request any
jobs that have a commandLine that includes a numeric argument.

subprocess.Popen requires that all arguments are strings though, so we
cannot just let the numeric arguments pass straight through.