Honor configured web port under gunicorn
The PTERO_SHELL_COMMAND_PORT is used by the flask application to
determine which port the web server should bind to when the flask app is
started directly.  The flask app is only started directly during
testing.

In production, we use gunicorn start the flask web application, so
gunicorn is responsible for binding to a port.

This change modifies the gunicorn command line in the production
Procfile to honor the PTERO_SHELL_COMMAND_PORT instead of listening on
80 by default.