Simplify handling of SIGTERM
The core issue appears to be how honcho sends SIGTERM to its children.
It used to be only to its children, but changed in later versions to be
recursive.  That caused our sigterm-wrapper to be bypassed.  This just
does the same conversion of SIGTERM to SIGINT within the same process.